+++ src/color.h
@@ -1,6 +1,8 @@
 #ifndef COLOR_H
 #define COLOR_H
 
-void HSBToRGB( unsigned int i_hue, unsigned int i_saturation, unsigned int i_brightness, unsigned int *o_red, unsigned int *o_green, unsigned int *o_blue ) ; 
+#include <stdint.h>
+
+void HSBToRGB( int16_t i_hue, uint8_t i_saturation, uint8_t i_brightness, uint8_t *o_red, uint8_t *o_green, uint8_t *o_blue ) ;
 
 #endif
